+++ setup.cfg
@@ -127,9 +127,9 @@ weaviate =
 graphdb = 
     SPARQLWrapper
 docstores =
-    farm-haystack[faiss,milvus1,weaviate,graphdb]
+    farm-haystack[faiss,milvus,weaviate,graphdb]
 docstores-gpu =
-    farm-haystack[faiss-gpu,milvus1,weaviate,graphdb]
+    farm-haystack[faiss-gpu,milvus,weaviate,graphdb]
 crawler = 
     selenium
     webdriver-manager
