+++ test/conftest.py
@@ -2,6 +2,7 @@ import subprocess
 import time
 from subprocess import run
 from sys import platform
+import os
 import gc
 import uuid
 import logging
@@ -15,13 +16,20 @@ import pytest
 import requests
 
 try:
+    from milvus import Milvus
+
+    milvus1 = True
+except ImportError:
+    milvus1 = False
+    from pymilvus import utility
+
+try:
     from elasticsearch import Elasticsearch
     from haystack.document_stores.elasticsearch import ElasticsearchDocumentStore
-    from milvus import Milvus
     import weaviate
 
     from haystack.document_stores.weaviate import WeaviateDocumentStore
-    from haystack.document_stores.milvus import MilvusDocumentStore
+    from haystack.document_stores import MilvusDocumentStore
     from haystack.document_stores.graphdb import GraphDBKnowledgeGraph
     from haystack.document_stores.faiss import FAISSDocumentStore
     from haystack.document_stores.sql import SQLDocumentStore
@@ -30,6 +38,7 @@ except (ImportError, ModuleNotFoundError) as ie:
     from haystack.utils.import_utils import _optional_component_not_installed
 
     _optional_component_not_installed("test", "test", ie)
+
 
 from haystack.document_stores import DeepsetCloudDocumentStore, InMemoryDocumentStore
 
@@ -107,13 +116,23 @@ def pytest_collection_modifyitems(config, items):
         # if the cli argument "--document_store_type" is used, we want to skip all tests that have markers of other docstores
         # Example: pytest -v test_document_store.py --document_store_type="memory" => skip all tests marked with "elasticsearch"
         document_store_types_to_run = config.getoption("--document_store_type")
+        document_store_types_to_run = document_store_types_to_run.split(",")
         keywords = []
+
+        if "milvus1" in document_store_types_to_run and not os.getenv("MILVUS1_ENABLED"):
+            document_store_types_to_run.remove("milvus1")
+            document_store_types_to_run.append("milvus")
+            if not milvus1:
+                raise Exception(
+                    "Milvus1 is enabled, but your pymilvus version only supports Milvus 2. Please select the correct pymilvus version."
+                )
+
         for i in item.keywords:
             if "-" in i:
                 keywords.extend(i.split("-"))
             else:
                 keywords.append(i)
-        for cur_doc_store in ["elasticsearch", "faiss", "sql", "memory", "milvus", "weaviate"]:
+        for cur_doc_store in ["elasticsearch", "faiss", "sql", "memory", "milvus1", "milvus", "weaviate"]:
             if cur_doc_store in keywords and cur_doc_store not in document_store_types_to_run:
                 skip_docstore = pytest.mark.skip(
                     reason=f'{cur_doc_store} is disabled. Enable via pytest --document_store_type="{cur_doc_store}"'
@@ -524,7 +543,7 @@ def ensure_ids_are_correct_uuids(docs: list, document_store: object) -> None:
             d["id"] = str(uuid.uuid4())
 
 
-@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus", "weaviate"])
+@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus1", "milvus", "weaviate"])
 def document_store_with_docs(request, test_docs_xs, tmp_path):
     embedding_dim = request.node.get_closest_marker("embedding_dim", pytest.mark.embedding_dim(768))
     document_store = get_document_store(
@@ -545,7 +564,7 @@ def document_store(request, tmp_path):
     document_store.delete_documents()
 
 
-@pytest.fixture(params=["memory", "faiss", "milvus", "elasticsearch"])
+@pytest.fixture(params=["memory", "faiss", "milvus1", "milvus", "elasticsearch"])
 def document_store_dot_product(request, tmp_path):
     embedding_dim = request.node.get_closest_marker("embedding_dim", pytest.mark.embedding_dim(768))
     document_store = get_document_store(
@@ -558,7 +577,7 @@ def document_store_dot_product(request, tmp_path):
     document_store.delete_documents()
 
 
-@pytest.fixture(params=["memory", "faiss", "milvus", "elasticsearch"])
+@pytest.fixture(params=["memory", "faiss", "milvus1", "milvus", "elasticsearch"])
 def document_store_dot_product_with_docs(request, test_docs_xs, tmp_path):
     embedding_dim = request.node.get_closest_marker("embedding_dim", pytest.mark.embedding_dim(768))
     document_store = get_document_store(
@@ -572,7 +591,7 @@ def document_store_dot_product_with_docs(request, test_docs_xs, tmp_path):
     document_store.delete_documents()
 
 
-@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus"])
+@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus1"])
 def document_store_dot_product_small(request, tmp_path):
     embedding_dim = request.node.get_closest_marker("embedding_dim", pytest.mark.embedding_dim(3))
     document_store = get_document_store(
@@ -585,7 +604,7 @@ def document_store_dot_product_small(request, tmp_path):
     document_store.delete_documents()
 
 
-@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus", "weaviate"])
+@pytest.fixture(params=["elasticsearch", "faiss", "memory", "milvus1", "milvus", "weaviate"])
 def document_store_small(request, tmp_path):
     embedding_dim = request.node.get_closest_marker("embedding_dim", pytest.mark.embedding_dim(3))
     document_store = get_document_store(
@@ -684,7 +703,7 @@ def get_document_store(
             isolation_level="AUTOCOMMIT",
         )
 
-    elif document_store_type == "milvus":
+    elif document_store_type == "milvus1":
         document_store = MilvusDocumentStore(
             embedding_dim=embedding_dim,
             sql_url=get_sql_url(tmp_path),
@@ -698,6 +717,17 @@ def get_document_store(
         for collection in collections:
             if collection.startswith(index):
                 document_store.milvus_server.drop_collection(collection)
+
+    elif document_store_type == "milvus":
+        document_store = MilvusDocumentStore(
+            embedding_dim=embedding_dim,
+            sql_url=get_sql_url(tmp_path),
+            return_embedding=True,
+            embedding_field=embedding_field,
+            index=index,
+            similarity=similarity,
+            isolation_level="AUTOCOMMIT",
+        )
 
     elif document_store_type == "weaviate":
         document_store = WeaviateDocumentStore(
