+++ test/test_pipeline.py
@@ -665,11 +665,11 @@ def test_save_to_deepset_cloud():
         ("dpr", "elasticsearch"),
         ("dpr", "faiss"),
         ("dpr", "memory"),
-        ("dpr", "milvus"),
+        ("dpr", "milvus1"),
         ("embedding", "elasticsearch"),
         ("embedding", "faiss"),
         ("embedding", "memory"),
-        ("embedding", "milvus"),
+        ("embedding", "milvus1"),
         ("elasticsearch", "elasticsearch"),
         ("es_filter_only", "elasticsearch"),
         ("tfidf", "memory"),
